optimizer: split converter
