+++ converter/optimizer/layerconverter/layerconverter.go
@@ -28,9 +28,9 @@ import (
 	"io"
 
 	"github.com/containerd/containerd/log"
+	"github.com/containerd/stargz-snapshotter/converter/optimizer/layer"
+	"github.com/containerd/stargz-snapshotter/converter/optimizer/logger"
 	"github.com/containerd/stargz-snapshotter/estargz"
-	"github.com/containerd/stargz-snapshotter/optimizer/layer"
-	"github.com/containerd/stargz-snapshotter/optimizer/logger"
 	"github.com/containerd/stargz-snapshotter/util/tempfiles"
 	regpkg "github.com/google/go-containerregistry/pkg/v1"
 	"github.com/google/go-containerregistry/pkg/v1/mutate"
